+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 + Fixed other entities render rotations from being 0'd out when using spoof rotations
 + Add more options to Animations module
 + Add Color options to Interface
++ Add "Fog" option to NoRender
 ```
 
 ---